[🚀 사이클2 - 미션 (예약 변경/취소와 에러 처리)] 아티(박지영) 미션 제출합니다. (#446)
* feat: 시간과 테마 중복 추가 및 삭제에 대한 예외 처리 추가

* feat: 서비스 정책에 따른 예약 예외처리 추가

* test: 서비스 정책에 따른 테스트 추가

* feat: 에러 응답 형식 적용 및 공통 에러 처리

* refactor: 커스텀 에러 비교로 테스트 수정

* feat: dto와 도메인에 null 값 검증 추가

* feat: @Validated 검증 예외 핸들러 추가

* refactor: 서버 기반 검증 및 예외 처리로 프론트 코드 변경

* feat: 예약 날짜/시간 변경 API 추가

* fix: 이미 지난 시간 삭제 예외처리 추가

* test: 예약 수정/취소 관련 테스트 추가

* feat: dto 검증 예외 처리 handler 추가

* feat: 예약 수정 API 프론트 연결

* docs: Readme update

* refactor: 시간 관련 도메인 Time으로 이름 통일

* refactor: 메서드명 오류 수정 및 중복 검증 제거

* fix: REST에 따라 생성 API에 헤더 정보 추가

* fix: 에러 응답 형식 수정

* fix: 변경사항에 따른 프론트 코드 수정

* chore: 자바 컨벤션에 따른 코드 정리

* docs: Readme update

+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,26 +1,48 @@
 package roomescape.service;
 
+import java.time.LocalDateTime;
 import java.util.List;
+
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import roomescape.domain.Reservation;
+import roomescape.domain.Time;
 import roomescape.dto.ReservationResponse;
+import roomescape.dto.ReservationUpdateRequest;
+import roomescape.exception.CustomException;
+import roomescape.exception.ErrorCode;
 import roomescape.repository.ReservationDao;
 import roomescape.dto.ReservationRequest;
+import roomescape.repository.TimeDao;
 
 @Service
 public class ReservationService {
     private final ReservationDao reservationDao;
+    private final TimeDao timeDao;
 
-    public ReservationService(ReservationDao reservationDao) {
+    public ReservationService(ReservationDao reservationDao, TimeDao timeDao) {
         this.reservationDao = reservationDao;
+        this.timeDao = timeDao;
     }
 
 
-    public ReservationResponse save(ReservationRequest request) {
-        Long id = reservationDao.save(request.name(), request.date(), request.timeId(), request.themeId());
+    public ReservationResponse save(LocalDateTime now, ReservationRequest request) {
+        Time reservationTime = timeDao.findById(request.timeId());
+        LocalDateTime time = LocalDateTime.of(request.date(), reservationTime.getStartAt());
+        validateDateAndTimeNotPast(now,time);
+
+        try{
+            Long id = reservationDao.save(request.name(), request.date(), request.timeId(), request.themeId());
+            Reservation reservation = reservationDao.findById(id);
+            return ReservationResponse.from(reservation);
+        } catch (DuplicateKeyException e){
+            throw new CustomException(ErrorCode.DUPLICATE_RESERVATION);
+        }
+    }
+
+    public ReservationResponse findById(long id) {
         Reservation reservation = reservationDao.findById(id);
         return ReservationResponse.from(reservation);
-
     }
 
     public List<ReservationResponse> findAllByName(String username) {
@@ -29,7 +51,29 @@ public class ReservationService {
                 .toList();
     }
 
-    public void delete(Long id) {
+    public void update(Long reservationId, LocalDateTime now, ReservationUpdateRequest request) {
+        try {
+            Time time = timeDao.findById(request.timeId());
+            LocalDateTime targetDateTime = LocalDateTime.of(request.targetDate(), time.getStartAt());
+            validateDateAndTimeNotPast(now,targetDateTime);
+            reservationDao.updateDateAndTimeById(reservationId,request.targetDate(), time.getId());
+        } catch (DuplicateKeyException e){
+            throw new CustomException(ErrorCode.DUPLICATE_RESERVATION);
+        }
+    }
+
+    public void delete(LocalDateTime now, Long id) {
+        Reservation reservation = reservationDao.findById(id);
+        LocalDateTime localDateTime = LocalDateTime.of(reservation.getDate(), reservation.getTime().getStartAt());
+        if (now.isAfter(localDateTime )) {
+            throw new CustomException(ErrorCode.UNALLOWED_DELETE_PAST_RESERVATION);
+        }
         reservationDao.delete(id);
+    }
+
+    private void validateDateAndTimeNotPast(LocalDateTime now, LocalDateTime reservationTime) {
+        if (now.isAfter(reservationTime)) {
+            throw new CustomException(ErrorCode.PAST_DATE_RESERVATION);
+        }
     }
 }
